% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activation_mish}
\alias{activation_mish}
\title{Mish}
\usage{
activation_mish(x)
}
\arguments{
\item{x}{A `Tensor`. Must be one of the following types: `float16`, `float32`, `float64`.
Returns: A `Tensor`. Has the same type as `x`.}
}
\value{
A `Tensor`. Has the same type as `x`.
}
\description{
Mish: A Self Regularized Non-Monotonic Neural Activation Function.
}
\details{
Computes mish activation: x * tanh(softplus(x))
See [Mish: A Self Regularized Non-Monotonic Neural Activation Function](https://arxiv.org/abs/1908.08681).
}
