% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activation_tanhshrink}
\alias{activation_tanhshrink}
\title{Tanhshrink}
\usage{
activation_tanhshrink(x)
}
\arguments{
\item{x}{x}
}
\value{
A `Tensor`. Has the same type as `features`.
}
\description{
Applies the element-wise function: x - tanh(x)
}
