% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{callback_time_stopping}
\alias{callback_time_stopping}
\title{Time Stopping}
\usage{
callback_time_stopping(seconds = 86400, verbose = 0)
}
\arguments{
\item{seconds}{maximum amount of time before stopping. Defaults to 86400 (1 day).}

\item{verbose}{verbosity mode. Defaults to 0.}
}
\value{
None
}
\description{
Time Stopping
}
\details{
Stop training when a specified amount of time has passed.
}
\examples{

\dontrun{
model \%>\% fit(
x_train, y_train,
batch_size = 128,
epochs = 4,
validation_split = 0.2,
verbose = 0,
callbacks = callback_time_stopping(seconds = 6, verbose = 1)
)
}



}
