% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ops.R
\name{crf_binary_score}
\alias{crf_binary_score}
\title{CRF binary score}
\usage{
crf_binary_score(tag_indices, sequence_lengths, transition_params)
}
\arguments{
\item{tag_indices}{A [batch_size, max_seq_len] matrix of tag indices.}

\item{sequence_lengths}{A [batch_size] vector of true sequence lengths.}

\item{transition_params}{A [num_tags, num_tags] matrix of binary potentials.}
}
\value{
binary_scores: A [batch_size] vector of binary scores.
}
\description{
Computes the binary scores of tag sequences.
}
