% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ops.R
\name{crf_decode}
\alias{crf_decode}
\title{CRF decode}
\usage{
crf_decode(potentials, transition_params, sequence_length)
}
\arguments{
\item{potentials}{A [batch_size, max_seq_len, num_tags] tensor of unary potentials.}

\item{transition_params}{A [num_tags, num_tags] matrix of binary potentials.}

\item{sequence_length}{A [batch_size] vector of true sequence lengths.}
}
\value{
decode_tags: A [batch_size, max_seq_len] matrix, with dtype `tf.int32`.
Contains the highest scoring tag indices. best_score: A [batch_size] vector,
containing the score of `decode_tags`.
}
\description{
Decode the highest scoring sequence of tags.
}
