% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ops.R
\name{crf_decode_forward}
\alias{crf_decode_forward}
\title{CRF decode forward}
\usage{
crf_decode_forward(inputs, state, transition_params, sequence_lengths)
}
\arguments{
\item{inputs}{A [batch_size, num_tags] matrix of unary potentials.}

\item{state}{A [batch_size, num_tags] matrix containing the previous step's score values.}

\item{transition_params}{A [num_tags, num_tags] matrix of binary potentials.}

\item{sequence_lengths}{A [batch_size] vector of true sequence lengths.}
}
\value{
backpointers: A [batch_size, num_tags] matrix of backpointers.
new_state: A [batch_size, num_tags] matrix of new score values.
}
\description{
Computes forward decoding in a linear-chain CRF.
}
