% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ops.R
\name{crf_log_norm}
\alias{crf_log_norm}
\title{CRF log norm}
\usage{
crf_log_norm(inputs, sequence_lengths, transition_params)
}
\arguments{
\item{inputs}{A [batch_size, max_seq_len, num_tags] tensor of unary potentials
to use as input to the CRF layer.}

\item{sequence_lengths}{A [batch_size] vector of true sequence lengths.}

\item{transition_params}{A [num_tags, num_tags] transition matrix.}
}
\value{
log_norm: A [batch_size] vector of normalizers for a CRF.
}
\description{
Computes the normalization for a CRF.
}
