% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ops.R
\name{crf_sequence_score}
\alias{crf_sequence_score}
\title{CRF sequence score}
\usage{
crf_sequence_score(inputs, tag_indices, sequence_lengths, transition_params)
}
\arguments{
\item{inputs}{A [batch_size, max_seq_len, num_tags] tensor of unary potentials
to use as input to the CRF layer.}

\item{tag_indices}{A [batch_size, max_seq_len] matrix of tag indices for which
we compute the unnormalized score.}

\item{sequence_lengths}{A [batch_size] vector of true sequence lengths.}

\item{transition_params}{A [num_tags, num_tags] transition matrix. Returns:}
}
\value{
sequence_scores: A [batch_size] vector of unnormalized sequence scores.
}
\description{
Computes the unnormalized score for a tag sequence.
}
