% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ops.R
\name{crf_unary_score}
\alias{crf_unary_score}
\title{CRF unary score}
\usage{
crf_unary_score(tag_indices, sequence_lengths, inputs)
}
\arguments{
\item{tag_indices}{A [batch_size, max_seq_len] matrix of tag indices.}

\item{sequence_lengths}{A [batch_size] vector of true sequence lengths.}

\item{inputs}{A [batch_size, max_seq_len, num_tags] tensor of unary potentials.}
}
\value{
unary_scores: A [batch_size] vector of unary scores.
}
\description{
Computes the unary scores of tag sequences.
}
