% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{decoder_base}
\alias{decoder_base}
\title{Base Decoder}
\usage{
decoder_base(object, cell, sampler, output_layer = NULL, ...)
}
\arguments{
\item{object}{Model or layer object}

\item{cell}{An RNNCell instance.}

\item{sampler}{A Sampler instance.}

\item{output_layer}{(Optional) An instance of tf$layers$Layer, i.e., tf$layers$Dense.
Optional layer to apply to the RNN output prior to storing the result or sampling.}

\item{...}{Other keyword arguments for layer creation.}
}
\value{
None
}
\description{
An RNN Decoder that is based on a Keras layer.
}
