% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{giou_loss}
\alias{giou_loss}
\title{Giou loss}
\usage{
giou_loss(y_true, y_pred, mode = "giou")
}
\arguments{
\item{y_true}{true targets tensor. The coordinates of the each bounding box in boxes are
encoded as [y_min, x_min, y_max, x_max].}

\item{y_pred}{predictions tensor. The coordinates of the each bounding box in boxes are
encoded as [y_min, x_min, y_max, x_max].}

\item{mode}{one of ['giou', 'iou'], decided to calculate GIoU or IoU loss.}
}
\value{
GIoU loss float `Tensor`.
}
\description{
Giou loss
}
