% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops_.R
\name{img_angles_to_projective_transforms}
\alias{img_angles_to_projective_transforms}
\title{Angles to projective transforms}
\usage{
img_angles_to_projective_transforms(
  angles,
  image_height,
  image_width,
  name = NULL
)
}
\arguments{
\item{angles}{A scalar angle to rotate all images by, or (for batches of images)
a vector with an angle to rotate each image in the batch. The rank must be statically
known (the shape is not `TensorShape(NULL)`.}

\item{image_height}{Height of the image(s) to be transformed.}

\item{image_width}{Width of the image(s) to be transformed.}

\item{name}{name of the op.}
}
\value{
A tensor of shape (num_images, 8). Projective transforms which can be given to `transform` op.
}
\description{
Returns projective transform(s) for the given angle(s).
}
