% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops_.R
\name{img_compose_transforms}
\alias{img_compose_transforms}
\title{Compose transforms}
\usage{
img_compose_transforms(transforms, name = NULL)
}
\arguments{
\item{transforms}{List of image projective transforms to be composed.
Each transform is length 8 (single transform) or shape (N, 8) (batched transforms).
The shapes of all inputs must be equal, and at least one input must be given.}

\item{name}{The name for the op.}
}
\value{
A composed transform tensor. When passed to `transform` op, equivalent to
applying each of the given transforms to the image in order.
}
\description{
Composes the transforms tensors.
}
