% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{img_equalize}
\alias{img_equalize}
\title{Equalize}
\usage{
img_equalize(image, bins = 256, name = NULL)
}
\arguments{
\item{image}{image}

\item{bins}{The number of bins in the histogram.}

\item{name}{The name of the op. Returns: Image(s) with the same type and shape as `images`, equalized.}
}
\value{
Image(s) with the same type and shape as `images`, equalized.
}
\description{
Equalize image(s)
}
