% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops_.R
\name{img_from_4D}
\alias{img_from_4D}
\title{From 4D image}
\usage{
img_from_4D(image, ndims)
}
\arguments{
\item{image}{4D tensor.}

\item{ndims}{The original rank of the image.}
}
\value{
`ndims`-D tensor with the same type.
}
\description{
Convert back to an image with `ndims` rank.
}
