% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{img_shear_y}
\alias{img_shear_y}
\title{Shear y-axis}
\usage{
img_shear_y(image, level, replace)
}
\arguments{
\item{image}{A 3D image Tensor.}

\item{level}{A float denoting shear element along x-axis}

\item{replace}{A one or three value 1D tensor to fill empty pixels.}
}
\value{
Transformed image along X or Y axis, with space outside image filled with replace.
}
\description{
Perform shear operation on an image (y-axis)
}
