% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops_.R
\name{img_to_4D}
\alias{img_to_4D}
\title{To 4D image}
\usage{
img_to_4D(image)
}
\arguments{
\item{image}{2/3/4D tensor.}
}
\value{
4D tensor with the same type.
}
\description{
Convert 2/3/4D image to 4D image.
}
\examples{

\dontrun{
img_to_4D(img)
}

}
