% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{img_translate_xy}
\alias{img_translate_xy}
\title{Translate xy dims}
\usage{
img_translate_xy(image, translate_to, replace)
}
\arguments{
\item{image}{A 3D image Tensor.}

\item{translate_to}{A 1D tensor to translate [x, y]}

\item{replace}{A one or three value 1D tensor to fill empty pixels.}
}
\value{
Translated image along X or Y axis, with space outside image
filled with replace. Raises: ValueError: if axis is neither 0 nor 1.
}
\description{
Translates image in X or Y dimension.
}
\section{Raises}{

ValueError: if axis is neither 0 nor 1.
}

