% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops_.R
\name{img_translations_to_projective_transforms}
\alias{img_translations_to_projective_transforms}
\title{Translations to projective transforms}
\usage{
img_translations_to_projective_transforms(translations, name = NULL)
}
\arguments{
\item{translations}{A 2-element list representing [dx, dy] or a matrix of 2-element
lists representing [dx, dy] to translate for each image (for a batch of images). The
rank must be statically known (the shape is not `TensorShape(NULL)`).}

\item{name}{The name of the op.}
}
\value{
A tensor of shape c(num_images, 8) projective transforms which can be given to `img_transform`.
}
\description{
Returns projective transform(s) for the given translation(s).
}
