% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops_.R
\name{img_unwrap}
\alias{img_unwrap}
\title{Uwrap}
\usage{
img_unwrap(image, replace)
}
\arguments{
\item{image}{image}

\item{replace}{a one or three value 1D tensor to fill empty pixels.}
}
\value{
a 3D image Tensor with 3 channels.
}
\description{
Unwraps an image produced by wrap.
}
\details{
Where there is a 0 in the last channel for every spatial position,
the rest of the three channels in that spatial dimension are grayed (set to 128).
Operations like translate and shear on a wrapped Tensor will leave 0s in empty
locations. Some transformations look at the intensity of values to do preprocessing,
and we want these empty pixels to assume the 'average' value, rather than pure black.
}
