% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops_.R
\name{img_wrap}
\alias{img_wrap}
\title{Wrap}
\usage{
img_wrap(image)
}
\arguments{
\item{image}{a 3D Image Tensor with 4 channels.}
}
\value{
'image' with an extra channel set to all 1s.
}
\description{
wrap an image array
}
