% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{layer_basic_decoder_output}
\alias{layer_basic_decoder_output}
\title{Basic decoder output}
\usage{
layer_basic_decoder_output(rnn_output, sample_id)
}
\arguments{
\item{rnn_output}{the output of RNN cell}

\item{sample_id}{the `id` of the sample}
}
\value{
None
}
\description{
Basic decoder output
}
