% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{layer_beam_search_decoder_output}
\alias{layer_beam_search_decoder_output}
\title{Beam Search Decoder Output}
\usage{
layer_beam_search_decoder_output(scores, predicted_ids, parent_ids)
}
\arguments{
\item{scores}{calculate the scores for each beam}

\item{predicted_ids}{The final prediction. A tensor of shape
`[batch_size, T, beam_width]` (or `[T, batch_size, beam_width]` if `output_time_major`
is `TRUE`). Beams are ordered from best to worst.}

\item{parent_ids}{The parent ids of shape `[max_time, batch_size, beam_width]`.}
}
\description{
Beam Search Decoder Output
}
