% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{layer_decoder}
\alias{layer_decoder}
\title{An RNN Decoder abstract interface object.}
\usage{
layer_decoder(...)
}
\arguments{
\item{...}{arguments to pass}
}
\value{
None
}
\description{
An RNN Decoder abstract interface object.
}
\details{
- inputs: (structure of) tensors and TensorArrays that is passed as input to the RNNCell
composing the decoder, at each time step.
- state: (structure of) tensors and TensorArrays that is passed to the RNNCell instance as the state.
- finished: boolean tensor telling whether each sequence in the batch is finished.
- training: boolean whether it should behave in training mode or in inference mode.
- outputs: Instance of BasicDecoderOutput. Result of the decoding, at each time step.
}
