% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{layer_final_beam_search_decoder_output}
\alias{layer_final_beam_search_decoder_output}
\title{Final Beam Search Decoder Output}
\usage{
layer_final_beam_search_decoder_output(
  predicted_ids,
  beam_search_decoder_output
)
}
\arguments{
\item{predicted_ids}{The final prediction. A tensor of shape `[batch_size, T, beam_width]`
(or `[T, batch_size, beam_width]` if `output_time_major` is TRUE). Beams are ordered from
best to worst.}

\item{beam_search_decoder_output}{An instance of `BeamSearchDecoderOutput` that describes
the state of the beam search.}
}
\description{
Final outputs returned by the beam search after all decoding is finished.
}
