% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{layer_gather_tree}
\alias{layer_gather_tree}
\title{Gather tree}
\usage{
layer_gather_tree(step_ids, parent_ids, max_sequence_lengths, end_token)
}
\arguments{
\item{step_ids}{requires the step id}

\item{parent_ids}{The parent ids of shape `[max_time, batch_size, beam_width]`.}

\item{max_sequence_lengths}{get max_sequence_length across all beams for each batch.}

\item{end_token}{`int32` scalar, the token that marks end of decoding.}
}
\value{
None
}
\description{
Gather tree
}
