% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{layer_hardmax}
\alias{layer_hardmax}
\title{Hardmax}
\usage{
layer_hardmax(logits, name = NULL)
}
\arguments{
\item{logits}{A batch tensor of logit values.}

\item{name}{Name to use when creating ops.}
}
\value{
A batched one-hot tensor.
}
\description{
Returns batched one-hot vectors.
}
\details{
The depth index containing the `1` is that of the maximum logit value.
}
