% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_maxout}
\alias{layer_maxout}
\title{Maxout layer}
\usage{
layer_maxout(object, num_units, axis = -1, ...)
}
\arguments{
\item{object}{Model or layer object}

\item{num_units}{Specifies how many features will remain after maxout in the axis dimension
(usually channel). This must be a factor of number of features.}

\item{axis}{The dimension where max pooling will be performed. Default is the last dimension.}

\item{...}{additional parameters to pass}
}
\value{
A tensor
}
\description{
Maxout layer
}
\details{
"Maxout Networks" Ian J. Goodfellow, David Warde-Farley, Mehdi Mirza,
Aaron Courville, Yoshua Bengio. https://arxiv.org/abs/1302.4389 Usually the operation
is performed in the filter/channel dimension. This can also be used after Dense layers
to reduce number of features.
}
