% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_poincare_normalize}
\alias{layer_poincare_normalize}
\title{Project into the Poincare ball with norm <= 1.0 - epsilon}
\usage{
layer_poincare_normalize(object, axis = 1, epsilon = 1e-05, ...)
}
\arguments{
\item{object}{Model or layer object}

\item{axis}{Axis along which to normalize.  A scalar or a vector of integers.}

\item{epsilon}{A small deviation from the edge of the unit sphere for numerical stability.}

\item{...}{additional parameters to pass}
}
\value{
A tensor
}
\description{
Project into the Poincare ball with norm <= 1.0 - epsilon
}
\details{
https://en.wikipedia.org/wiki/Poincare_ball_model Used in Poincare Embeddings
for Learning Hierarchical Representations Maximilian Nickel, Douwe Kiela
https://arxiv.org/pdf/1705.08039.pdf For a 1-D tensor with axis = 0, computes
}
