% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{layer_sampler}
\alias{layer_sampler}
\title{Sampler}
\usage{
layer_sampler(...)
}
\arguments{
\item{...}{parametr to pass batch_size, initialize, next_inputs, sample, sample_ids_dtype, sample_ids_shape}
}
\value{
None
}
\description{
Interface for implementing sampling in seq2seq decoders.
}
