% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{lifted_struct_loss}
\alias{lifted_struct_loss}
\title{Lifted structured loss}
\usage{
lifted_struct_loss(labels, embeddings, margin = 1)
}
\arguments{
\item{labels}{1-D tf.int32 Tensor with shape [batch_size] of multiclass integer labels.}

\item{embeddings}{2-D float Tensor of embedding vectors. Embeddings should not be l2 normalized.}

\item{margin}{Float, margin term in the loss definition.}
}
\value{
lifted_loss: tf.float32 scalar.
}
\description{
Computes the lifted structured loss.
}
