% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_contrastive}
\alias{loss_contrastive}
\title{Contrastive loss}
\usage{
loss_contrastive(y_true, y_pred, margin = 1)
}
\arguments{
\item{y_true}{1-D integer Tensor with shape [batch_size] of binary labels indicating positive vs negative pair.}

\item{y_pred}{1-D float Tensor with shape [batch_size] of distances between two embedding matrices.}

\item{margin}{margin term in the loss definition.}
}
\value{
contrastive_loss: 1-D float `Tensor` with shape [batch_size].
}
\description{
Computes the contrastive loss between `y_true` and `y_pred`.
}
\details{
This loss encourages the embedding to be close to each other for
the samples of the same label and the embedding to be far apart at least
by the margin constant for the samples of different labels.
The euclidean distances `y_pred` between two embedding matrices
`a` and `b` with shape [batch_size, hidden_size] can be computed
as follows: ```python
# y_pred = `\\sqrt` (`\\sum_i` (a[:, i] - b[:, i])^2)
y_pred = tf$linalg.norm(a - b, axis=1)
``` See: http://yann.lecun.com/exdb/publis/pdf/hadsell-chopra-lecun-06.pdf
}
