% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{loss_sequence}
\alias{loss_sequence}
\title{Weighted cross-entropy loss for a sequence of logits.

... A list of parameters}
\usage{
loss_sequence(...)
}
\value{
None
}
\description{
Weighted cross-entropy loss for a sequence of logits.

... A list of parameters
}
