% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_triplet_hard}
\alias{loss_triplet_hard}
\title{Triplet hard loss}
\usage{
loss_triplet_hard(margin = 1, soft = FALSE, name = NULL, ...)
}
\arguments{
\item{margin}{Float, margin term in the loss definition. Default value is 1.0.}

\item{soft}{Boolean, if set, use the soft margin version. Default value is False.}

\item{name}{Optional name for the op.}

\item{...}{additional arguments to pass}
}
\value{
triplet_loss: float scalar with dtype of y_pred.
}
\description{
Computes the triplet loss with hard negative and hard positive mining.
}
