% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_triplet_semihard}
\alias{loss_triplet_semihard}
\title{Triplet semihard loss}
\usage{
loss_triplet_semihard(y_true, y_pred, margin = 1)
}
\arguments{
\item{y_true}{1-D integer Tensor with shape [batch_size] of multiclass integer labels.}

\item{y_pred}{2-D float Tensor of embedding vectors. Embeddings should be l2 normalized.}

\item{margin}{Float, margin term in the loss definition.}
}
\value{
triplet_loss: float scalar with dtype of y_pred.
}
\description{
Computes the triplet loss with semi-hard negative mining.
}
