% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_hamming_distance}
\alias{metric_hamming_distance}
\title{Hamming distance}
\usage{
metric_hamming_distance(actuals, predictions)
}
\arguments{
\item{actuals}{actual target value}

\item{predictions}{predicted value}
}
\value{
hamming distance: float
}
\description{
Computes hamming distance.
}
\details{
Hamming distance is for comparing two binary strings.
It is the number of bit positions in which two bits
are different.
}
\examples{

\dontrun{

actuals = tf$constant(as.integer(c(1, 1, 0, 0, 1, 0, 1, 0, 0, 1)), dtype=tf$int32)
predictions = tf$constant(as.integer(c(1, 0, 0, 0, 1, 0, 0, 1, 0, 1)),dtype=tf$int32)
result = metric_hamming_distance(actuals, predictions)
paste('Hamming distance: ', result$numpy())

}

}
