% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_objects.R
\name{register_custom_kernels}
\alias{register_custom_kernels}
\title{Register custom kernels}
\usage{
register_custom_kernels(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Register custom kernels
}
