% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_objects.R
\name{register_keras_objects}
\alias{register_keras_objects}
\title{Register keras objects}
\usage{
register_keras_objects(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
Register keras objects
}
