% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2seq.R
\name{sampler_training}
\alias{sampler_training}
\title{A Sampler for use during training.}
\usage{
sampler_training(time_major = FALSE)
}
\arguments{
\item{time_major}{bool. Whether the tensors in inputs are time major.
If `FALSE` (default), they are assumed to be batch major.}
}
\value{
None
}
\description{
Only reads inputs.
}
