% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops_.R
\name{sharpness}
\alias{sharpness}
\title{Sharpness}
\usage{
sharpness(image, factor)
}
\arguments{
\item{image}{an image}

\item{factor}{A floating point value or Tensor above 0.0.}
}
\value{
Image(s) with the same type and shape as `images`, sharper.
}
\description{
Change sharpness of image(s)
}
