% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{sigmoid_focal_crossentropy_loss}
\alias{sigmoid_focal_crossentropy_loss}
\title{Sigmoid focal crossentropy loss}
\usage{
sigmoid_focal_crossentropy_loss(
  y_true,
  y_pred,
  alpha = 0.25,
  gamma = 2,
  from_logits = FALSE
)
}
\arguments{
\item{y_true}{true targets tensor.}

\item{y_pred}{predictions tensor.}

\item{alpha}{balancing factor.}

\item{gamma}{modulating factor.}

\item{from_logits}{If logits are provided then convert the predictions into probabilities}
}
\value{
Weighted loss float `Tensor`. If `reduction` is `NONE`,this has the same shape as `y_true`;
otherwise, it is scalar.
}
\description{
Sigmoid focal crossentropy loss
}
