% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tfaddons_version}
\alias{tfaddons_version}
\title{Version of TensorFlow SIG Addons}
\usage{
tfaddons_version()
}
\value{
prints the version.
}
\description{
Get the current version of TensorFlow SIG Addons
}
