% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ops.R
\name{viterbi_decode}
\alias{viterbi_decode}
\title{Viterbi decode}
\usage{
viterbi_decode(score, transition_params)
}
\arguments{
\item{score}{A [seq_len, num_tags] matrix of unary potentials.}

\item{transition_params}{A [num_tags, num_tags] matrix of binary potentials.}
}
\value{
viterbi: A [seq_len] list of integers containing the highest scoring tag indices.
viterbi_score: A float containing the score for the Viterbi sequence.
}
\description{
Decode the highest scoring sequence of tags outside of TensorFlow.
}
\details{
This should only be used at test time.
}
