% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{callback_time_stopping}
\alias{callback_time_stopping}
\title{Time Stopping}
\usage{
callback_time_stopping(seconds = 86400, verbose = 0)
}
\arguments{
\item{seconds}{maximum amount of time before stopping. Defaults to 86400 (1 day).}

\item{verbose}{verbosity mode. Defaults to 0.}
}
\value{
None
}
\description{
Time Stopping
}
\details{
Stop training when a specified amount of time has passed.
}
