% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ops.R
\name{img_euclidean_dist_transform}
\alias{img_euclidean_dist_transform}
\title{Euclidean dist transform}
\usage{
img_euclidean_dist_transform(images, dtype = tf$float32, name = NULL)
}
\arguments{
\item{images}{A tensor of shape (num_images, num_rows, num_columns, 1) (NHWC),
or (num_rows, num_columns, 1) (HWC) or (num_rows, num_columns) (HW).}

\item{dtype}{DType of the output tensor.}

\item{name}{The name of the op.}
}
\value{
Image(s) with the type `dtype` and same shape as `images`, with the
transform applied. If a tensor of all ones is given as input, the output tensor
will be filled with the max value of the `dtype`.
}
\description{
Applies euclidean distance transform(s) to the image(s).
}
\section{Raises}{

TypeError: If `image` is not tf.uint8, or `dtype` is not floating point.
ValueError: If `image` more than one channel, or `image` is not of rank between 2 and 4.
}

