% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_sparsemax}
\alias{loss_sparsemax}
\title{Sparsemax loss}
\usage{
loss_sparsemax(logits, sparsemax, labels, name = NULL)
}
\arguments{
\item{logits}{A `Tensor`. Must be one of the following types: `float32`, `float64`.}

\item{sparsemax}{A `Tensor`. Must have the same type as `logits`.}

\item{labels}{A `Tensor`. Must have the same type as `logits`.}

\item{name}{A name for the operation (optional). Returns: A `Tensor`. Has the same type as `logits`.}
}
\value{
A `Tensor`. Has the same type as `logits`.
}
\description{
Sparsemax loss function [1].
}
\details{
Computes the generalized multi-label classification loss for the sparsemax
function. The implementation is a reformulation of the original loss
function such that it uses the sparsemax properbility output instead of the
internal au variable. However, the output is identical to the original
loss function. [1]: https://arxiv.org/abs/1602.02068
}
