% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_triplet_hard}
\alias{loss_triplet_hard}
\title{Triplet hard loss}
\usage{
loss_triplet_hard(y_true, y_pred, margin = 1, soft = FALSE)
}
\arguments{
\item{y_true}{1-D integer `Tensor` with shape [batch_size] of
multiclass integer labels.}

\item{y_pred}{2-D float `Tensor` of embedding vectors. Embeddings should
be l2 normalized.}

\item{margin}{Float, margin term in the loss definition.}

\item{soft}{Boolean, if set, use the soft margin version.}
}
\value{
triplet_loss: float scalar with dtype of y_pred.
}
\description{
Computes the triplet loss with hard negative and hard positive mining.
}
