% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_triplet_semihard}
\alias{loss_triplet_semihard}
\title{Triplet semihard loss}
\usage{
loss_triplet_semihard(margin = 1, name = NULL, ...)
}
\arguments{
\item{margin}{Float, margin term in the loss definition. Default value is 1.0.}

\item{name}{Optional name for the op.}

\item{...}{additional arguments to pass}
}
\value{
triplet_loss: float scalar with dtype of y_pred.
}
\description{
Computes the triplet loss with semi-hard negative mining.
}
